Move fetch_redash() to poucave.utils
