+++ poucave/utils.py
@@ -1,6 +1,8 @@
 from datetime import datetime, timedelta
 from typing import Dict, Any, Tuple, Optional
 
+import aiohttp
+
 
 class Cache:
     def __init__(self):
@@ -24,3 +26,19 @@ class Cache:
         except KeyError:
             # Unknown key.
             return None
+
+
+REDASH_URI = "https://sql.telemetry.mozilla.org/api/queries/{}/results.json?api_key={}"
+
+
+async def fetch_redash(query_id, api_key):
+    redash_uri = REDASH_URI.format(query_id, api_key)
+
+    async with aiohttp.ClientSession() as session:
+        async with session.get(redash_uri) as response:
+            body = await response.json()
+
+    query_result = body["query_result"]
+    data = query_result["data"]
+    rows = data["rows"]
+    return rows
